:truck: Fix: 파일 구조 변경 추가
