+++ client/src/components/contents/Content.tsx
@@ -1,8 +1,8 @@
 import { useEffect } from 'react';
 import { useParams } from 'react-router-dom';
-import CommentSection from '@/components/comments/CommentSection';
-import ContentDetail from '@/components/contents/ContentDetail';
-import RecommendContent from '@/components/contents/RecommendContent';
+import CommentSection from '@/components/contents/comments/CommentSection';
+import ContentDetail from '@/components/contents/contentDetail/ContentDetail';
+import RecommendContent from '@/components/contents/recommendContent/RecommendContent';
 
 function Content() {
   const { id = '' } = useParams<{ id?: string }>();
