+++ client/src/components/contents/contentDetail/ContentDetail.tsx
@@ -5,10 +5,10 @@ import styled, { keyframes } from 'styled-components';
 import { GetDataDetail, GetUser } from '@/api/api';
 import DeleteMediaBtn from '@/components/admin/DeleteMediaBtn';
 import PatchMediaBtn from '@/components/admin/PatchMediaBtn';
-import Bookmark from '@/components/contents/Bookmark';
-import ContentDetailLoading from '@/components/contents/ContentDetailLoading';
-import Recommend from '@/components/contents/Recommend';
-import ReportButton from '@/components/contents/ReportButton';
+import ReportButton from '@/components/button/ReportButton';
+import Bookmark from '@/components/contents/bookmark/Bookmark';
+import ContentDetailLoading from '@/components/contents/contentDetail/ContentDetailLoading';
+import Recommend from '@/components/contents/recommend/Recommend';
 import Tag from '@/components/contents/Tag';
 import useMediaQuery from '@/hooks/useMediaQuery';
 import Error from '@/pages/Error';
