+++ client/src/components/contents/recommendContent/RecommendContent.tsx
@@ -3,8 +3,8 @@ import { AxiosError } from 'axios';
 import { useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
 import { GetDataDetail, GetFilterdData } from '@/api/api';
-import { RecommendContentLoading } from '@/components/contents/RecommendContentLoading';
 import ItemCard from '@/components/card/ItemCard';
+import { RecommendContentLoading } from '@/components/contents/recommendContent/RecommendContentLoading';
 import { ContentData } from '@/types/types';
 
 const RecommendContent = ({ contentId }: { contentId: string }) => {
